Feature 1791108 - Invidiual Page Locking. Implemented by adding TopicIsReadOnly property to the ContentProviderChain and exposed it to WikiTalk as TopicInfo.IsLocked.
Revised the buitin _NormalBorders to include a check for namespace.HasManageNamespacePermission and if true display links in the Right Border to the default administration page and to a new page admin/TopicLocks.aspx

TopicLocks.aspx is built by revising a copy of LastModified.aspx so that is shows the lock status of all topics in a namespace. If the user has the ManageNamespacePermission then it also displays a button for each topic that allows that topic to be locked/unlocked based on the topic lock status.

No unit tests were included as the functionality is dependent upon a bit in the file system implementation that is not available in the MockContentStore (I could not determine how to implement such a property in a meaningful way. Changes have been made throughout unit test ContentProviderChain up to that last point.) 

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@862 662ad2f3-a61f-0410-a8b9-97bb17fbd72e