Feature 1403950 - Automatic Table of Contents
Modified the TopicParser to read the header lines in a topic into a retrievable variable. That retrieved header information is then used to generate the table of contents for a topic. The TOC can be formatted for either bullets or numeric outlines. The TOC has links to the headers in the original topic, which have had anchors added (not to interfere with the anchors already generated for non-existant topics). 

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@868 662ad2f3-a61f-0410-a8b9-97bb17fbd72e