Feature Feature 1891349 - Add favicon capability.
Added the capability to output a favicon on FlexWiki web pages. Added a web application configuration variable DisableFavicon with a default value of false. If the favicon is not present or DisableFavicon is true then the webpages do not output the link. Added the favicon.ico file to the main web site and the admin directory


git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@925 662ad2f3-a61f-0410-a8b9-97bb17fbd72e