Bug 1892979 - SqlStore Topic Locking Issues. Fixes a bug in how IsTopicReadOnly works in the SqlStore and implements TopicLock/Unlock for the SqlStore.
NOTE: Current SqlStore implementations must also download the tools distribution to get a copy of an update script FlexWikiSqlStoreObjectUpdateScript.sql and run that script against their current store to add 2 new Stored Procedures.

Making an editorial change to DatabaseHelper.cs to force a rebuild of the tools. Revising a script file was not sufficient to cause the lib to be rebuilt.

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@930 662ad2f3-a61f-0410-a8b9-97bb17fbd72e