Bug 1892979 - SqlStore Topic Locking Issues. Fixes a bug in how IsTopicReadOnly works in the SqlStore and implements TopicLock/Unlock for the SqlStore.
NOTE: Current SqlStore implementations must also download the tools distribution to get a copy of an update script FlexWikiSqlStoreObjectUpdateScript.sql and run that script against their current store to add 2 new Stored Procedures.

Changed references to SqlStoreManagerLib.dll in the update csproj files to try to force a rebuild of the tools to use the revised scripts. The issue appears to be that the dll is not properly versioned in the build system

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@932 662ad2f3-a61f-0410-a8b9-97bb17fbd72e