Feature 1915966 - Allow CAPTCHA output image to be configurable
Added configuration variables for image HatchStyle, WarpMode, ForegroundColor, BackgroundColor and ObscuringLines. Each variable has a default value as per the original program and defaults will be used when there is an error reading or using the supplied value in the variable. The allowed values have been listed in the flexwiki.config.template file for each variable.

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@991 662ad2f3-a61f-0410-a8b9-97bb17fbd72e