Feature 1921373 - Add an Admin Page to Delete a Topic
Implemented a page to delete a topic and all its history. Modified the DeleteTopic method to include a boolean to determine whether or not history .awiki files were to be removed at the same time as the topic. This was implemented for both the FileSystem Store and the SQLStore.

Access to the Delete Topic and History functionality is restricted to users with ManageNamespacePermission. 

Modified existing unit tests to include the new boolean parameter, where the history file is not deleted. Created unit tests for the FileSystemStore and the SqlStore where the history files are deleted when the topic is deleted. Added a method to implement the GetSqlTopicArchiveInfos stored procedure for the SqlProvider/MockDatabase.

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@1000 662ad2f3-a61f-0410-a8b9-97bb17fbd72e