Feature 1924298 - Expose Registered User stats when available
Added an IMembership interface to the FlexWiki engine. Implemented an instance of FlexWikiWebMembership in association with FlexWikiWebApplication. IMembership calls for 2 properties to allow counting number of registered users and the number of currently only users. These 2 properties are exposed via the Federation object. When no membership provider is used these properties return 0.

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@1003 662ad2f3-a61f-0410-a8b9-97bb17fbd72e