Feature 1928955 - MessagePost improvements
Convert sequences like @@WikiTalk@@ or {{IncludeFile}} to be escaped by enclosing in a pair of double -quotes as shown; ""@@WikiTalk@@"" and ""{{IncludeFile}}""; for MessagePost output only.

Enabled the Preview function for MessagePost.

Moved javascript code for thread management to WikiThreads.js and include in output of default.aspx.

Moved Wikitalk for forum support to a library called ForumLibrary. Created a new namespace called WikiTalkLibrary and referenced this new namespace in flexwiki.config.template. This library has been added to the svn controlled files and will be included in the distribution. ReflectionLibrary, BlogSupportLibrary and PictureGalleryLibrary are also included in this namespace.

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@1010 662ad2f3-a61f-0410-a8b9-97bb17fbd72e