Feature 2116466 - Enable New Parser functionality at Topic level
Added a new FlexWikiApplicationConfiguration item <DisableNewParser> with a default value of false. When this is true all new parser functionality is disabled. 

A Topic level property "DisplayUsingNewParser" when set to true will employ the new parser providing <DisableNewParser> is false.

NOTE: Functionality for OrderedListForContinuation and OrderedListContinued is not complete in this build, but all other new parser functionality is complete.

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@1060 662ad2f3-a61f-0410-a8b9-97bb17fbd72e