Feature 2123292 - Improve Keyword Use
Implemented 2 new types, KeywordInfo and KeywordInfoArray, that are exposed via a WikiTalk NamespaceInfo. KeywordInfoArray holds an ArrayList of KeywordInfo that can be used to create a tag cloud. This has an entry for each keyword used in the namespace and a count of the number of times it is referenced.

A WikiTalk script will be deployed at flexwiki.com in the topic KeywordReferences, that will display the Tag Cloud for a namespace, or a list of topics in that namespace referenced by the selected keyword. 

The _NormalBorders in flexwiki.com will also be modified to display a list of reference links when a topic includes a Keywords property.

Unit tests were added for the new functionality, and also for TopicInfo.Keywords and TopicInfo.KeywordsList

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@1064 662ad2f3-a61f-0410-a8b9-97bb17fbd72e