Bug 2175394 - Improve XSS and XSRF protections
Deleted the original session id cookie when a successful login occurs, which causes a new cookie to be created.
Add configuration variables XsrfProtectionMessagePostTimeout and XsrfProtectionWikiEditTimeout to flexwiki.config, documented in flexwiki.config.template. Default timeout periods are 10 and 15 minutes respectively.

Added a nonce to the forms in WikiEdit and MessagePost and generate a unique cookie for each form instance. The cookie and the nonce values must match for the save to occur. This guarantees that the save action came from a browser that had previously communicated with the flexwiki hosted site and that it was not the result of activity from a malicious site, or input from a bot.

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@1078 662ad2f3-a61f-0410-a8b9-97bb17fbd72e