Feature 2180669 - Encrypt values of XSRF protection
Added passphrases for encrypting the nonce and cookie used for xsrf protection. The passphrases may be 32 or 16-bytes in length. There are 16-byte default passphrases to ensure simple transition. Modified the WikiEdit and MessagePost xsrf routines to use encryption and decryption.

Added a unit test for encryption and decryption longer passphrases.

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@1083 662ad2f3-a61f-0410-a8b9-97bb17fbd72e