Feature 2187942 - Add Atom Feed output at topic level
Implemented a new page type AtomFeed.aspx, which is an Atom 1.0 feed output, intended for use with a FlexWiki-based blog. The blog page can have a single entry or multiple entries.

Added a new method to the WikiTalk Utility class to generate a UUID (GUID) for use with an atom feed.

Added a check in default.aspx for the existence of a property named CreateAtomFeedLink and if present with a value of true it will generate a link to the atom feed for this topic. 

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@1087 662ad2f3-a61f-0410-a8b9-97bb17fbd72e