Feature 2192011 - Add Logo, favicon and tags to AtomFeed
Output the <logo> tag in the AtomFeed, based on a property FeedLogo in the base topic.
Output the <icon> tag in the AtomFeed, based on a property FeedIcon in the base topic.
Output the <category> tag in the entry section of the AtomFeed, based on the Keywords property in the entry topic.

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@1093 662ad2f3-a61f-0410-a8b9-97bb17fbd72e