Feature 2242777 - Implement PasswordReset without sending Email
Created a new page PasswordReset.aspx that implements a wizard control to verify a user using their security question and then update their password to a new password of their choosing without send an email.

Added a configuration item PasswordResetUrl to flexwiki.config with a default of '~/RecoverPassword.aspx' so that the login control on Login.aspx can point to either the standard password recovery facility or new custom password reset.

git-svn-id: https://flexwiki.svn.sourceforge.net/svnroot/flexwiki/flexwikicore/trunk@1111 662ad2f3-a61f-0410-a8b9-97bb17fbd72e